requirements.txt
requests
streamlit
python-dotenv
openai
imbalanced-learn
pandas
matplotlib
scikit-learn
numpy
seaborn
